[NCL-4820] Support for temporary builds for Gradle
