+++ repour/adjust/gradle_provider.py
@@ -19,7 +19,7 @@ INIT_SCRIPT_FILE_NAME = "analyzer-init.gradle"
 MANIPULATION_FILE_NAME = "manipulation.json"
 
 
-def get_gradle_provider(init_file_path, default_parameters):
+def get_gradle_provider(init_file_path, default_parameters, specific_indy_group=None, timestamp=None):
 
     @asyncio.coroutine
     def adjust(work_dir, extra_adjust_parameters, adjust_result):
@@ -28,6 +28,16 @@ def get_gradle_provider(init_file_path, default_parameters):
         if not os.path.exists(init_file_path):
             raise Exception(
                 "The Gradle init file '{}' does not exist - are you sure you provided the correct path in configuration?".format(init_file_path))
+
+        temp_build_parameters = []
+
+        if timestamp:
+            temp_build_parameters.append(
+                "-DversionIncrementalSuffix=" + timestamp + "-redhat")
+
+        if specific_indy_group:
+            temp_build_parameters.append(
+                "-DrestRepositoryGroup=" + specific_indy_group)
 
         logger.info("Adjusting in {}".format(work_dir))
         logger.info("Copying Gradle init file from '{}'".format(init_file_path))
@@ -46,7 +56,7 @@ def get_gradle_provider(init_file_path, default_parameters):
         )
 
         cmd = ["./gradlew", "--info", "--console", "plain", "--no-daemon", "--stacktrace",
-               "--init-script", INIT_SCRIPT_FILE_NAME, "generateAlignmentMetadata"] + default_parameters
+               "--init-script", INIT_SCRIPT_FILE_NAME, "generateAlignmentMetadata"] + default_parameters + temp_build_parameters
 
         result = yield from process_provider.get_process_provider(EXECUTION_NAME,
                                                                   cmd,
